Use decimal instead of float for internal calculations all throughout the library

+++ src/ctrader_api_client/events/types.py
@@ -30,8 +30,8 @@ class SpotEvent:
 
     account_id: int
     symbol_id: int
-    bid: float | None
-    ask: float | None
+    bid: Decimal | None
+    ask: Decimal | None
     trendbar: Trendbar | None
     timestamp: datetime
